requirements.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
et_xmlfile==2.0.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
idna==3.11
numpy==2.3.4
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
